Add engine aware mixin to factor out engine stuff
